requirements.txt
click==8.1.7
dbus_python==1.3.2
git+https://github.com/Fabric-Development/fabric.git
loguru==0.7.2
psutil==6.1.0
pycairo==1.27.0
PyGObject==3.50.0
PyGObject-stubs==2.10.0
pyjson5==1.6.8
pytomlpp==1.0.13
rlottie-python==1.3.6
setproctitle==1.3.4
